# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: AutoCare Hub API
  version: 1.0.0
  description: >
    API para gerenciamento de clientes, veículos, serviços, peças, ordens de
    serviço, orçamentos, aprovações e acompanhamento de status de uma oficina
    mecânica.
servers:
  - url: http://localhost:8080
    description: Ambiente local
security:
  - bearerAuth: [ ]
...
paths:
...
  /api/v1/parts:
    post:
      tags:
        - Parts
      summary: Cria uma peça ou insumo
      operationId: createPart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePartRequest'
            examples:
              peca:
                summary: Peça com estoque inicial
                value:
                  name: Sensor de oxigenio universal
                  description: Sensor lambda universal para diagnostico de mistura ar-combustivel.
                  sku: MOT-SEN-OXI01
                  category: Motor
                  subcategory: Injecao
                  brand: Bosch
                  unitPrice: 320.00
                  costPrice: 210.00
                  stockQuantity: 12
                  minimumStock: 3
      responses:
        '201':
          description: Peça criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    get:
      tags:
        - Parts
      summary: Lista peças e insumos
      operationId: listParts
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: active
          in: query
          required: false
          schema:
            type: boolean
        - name: lowStock
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Peças encontradas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/parts/{partId}:
    get:
      tags:
        - Parts
      summary: Busca uma peça por id
      operationId: getPartById
      parameters:
        - $ref: '#/components/parameters/PartIdPathParam'
      responses:
        '200':
          description: Peça encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Parts
      summary: Atualiza uma peça ou insumo
      operationId: updatePart
      parameters:
        - $ref: '#/components/parameters/PartIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePartRequest'
            examples:
              atualizarPeca:
                summary: Atualiza peça mantendo ativa
                value:
                  name: Sensor de oxigenio universal
                  description: Sensor lambda universal para diagnostico de mistura ar-combustivel.
                  sku: MOT-SEN-OXI01
                  category: Motor
                  subcategory: Injecao
                  brand: Bosch
                  unitPrice: 340.00
                  costPrice: 220.00
                  stockQuantity: 12
                  minimumStock: 3
                  active: true
      responses:
        '200':
          description: Peça atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Parts
      summary: Remove logicamente uma peça ou insumo
      operationId: deletePart
      parameters:
        - $ref: '#/components/parameters/PartIdPathParam'
      responses:
        '204':
          description: Peça removida
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/parts/{partId}/stock:
    patch:
      tags:
        - Parts
      summary: Atualiza o estoque de uma peça ou insumo
      operationId: updatePartStock
      parameters:
        - $ref: '#/components/parameters/PartIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePartStockRequest'
            examples:
              ajusteEstoque:
                summary: Ajuste direto da quantidade total em estoque
                value:
                  stockQuantity: 25
      responses:
        '200':
          description: Estoque atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/parts/{partId}/stock-movement:
    patch:
      tags:
        - Parts
      summary: Registra entrada, saída ou venda isolada de uma peça
      operationId: registerStockMovement
      parameters:
        - $ref: '#/components/parameters/PartIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockMovementRequest'
            examples:
              entradaCompra:
                summary: Entrada de compra de fornecedor
                value:
                  type: ENTRY
                  quantity: 10
                  unitCost: 210.00
                  unitPrice: 320.00
                  reason: Compra de reposicao de estoque
              saidaInterna:
                summary: Saída manual de estoque
                value:
                  type: EXIT
                  quantity: 1
                  reason: Ajuste por perda operacional
      responses:
        '200':
          description: Movimento registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/parts/{partId}/reservation:
    patch:
      tags:
        - Parts
      summary: Configura o prazo de reserva de estoque para orçamentos
      operationId: configurePartReservation
      parameters:
        - $ref: '#/components/parameters/PartIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigurePartReservationRequest'
            examples:
              prazoReserva:
                summary: Prazo de reserva usado em orçamentos
                value:
                  reservationDays: 3
      responses:
        '200':
          description: Prazo de reserva configurado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/parts/{partId}/reserve:
    patch:
      tags:
        - Parts
      summary: Bloqueia quantidade de uma peça para um orçamento
      operationId: reservePart
      parameters:
        - $ref: '#/components/parameters/PartIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartQuantityRequest'
            examples:
              reservar:
                summary: Reserva quantidade para orçamento
                value:
                  quantity: 2
      responses:
        '200':
          description: Estoque reservado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/parts/{partId}/release-reservation:
    patch:
      tags:
        - Parts
      summary: Libera quantidade previamente reservada
      operationId: releaseReservation
      parameters:
        - $ref: '#/components/parameters/PartIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartQuantityRequest'
            examples:
              liberarReserva:
                summary: Libera quantidade reservada
                value:
                  quantity: 1
      responses:
        '200':
          description: Reserva liberada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/parts/{partId}/commit-reservation:
    patch:
      tags:
        - Parts
      summary: Converte reserva em baixa definitiva de estoque
      operationId: commitReservation
      parameters:
        - $ref: '#/components/parameters/PartIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommitPartReservationRequest'
            examples:
              baixarReserva:
                summary: Converte reserva em baixa de estoque
                value:
                  quantity: 1
                  reason: Peca utilizada na execucao da OS
      responses:
        '200':
          description: Reserva baixada do estoque
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    SizeParam:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    PartIdPathParam:
      name: partId
      in: path
      required: true
      example: 40000000-0000-0000-0000-000000000005
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Requisição invalida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2026-06-20T17:00:00-03:00'
            status: 400
            error: Bad Request
            message: Invalid request body
            path: /api/v1/service-orders
            details:
              - customerDocument: must not be blank
    Unauthorized:
      description: Credenciais ausentes ou invalidas
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2026-06-20T17:00:00-03:00'
            status: 401
            error: Unauthorized
            message: Missing or invalid credentials
            path: /api/v1/customers
            details: [ ]
...
    NotFound:
      description: Recurso nao encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2026-06-20T17:00:00-03:00'
            status: 404
            error: Not Found
            message: Resource not found
            path: /api/v1/customers/00000000-0000-0000-0000-000000000000
            details: [ ]
...
  schemas:
...
    CreatePartRequest:
      type: object
      required:
        - name
        - description
        - sku
        - category
        - brand
        - unitPrice
        - stockQuantity
        - minimumStock
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
        description:
          type: string
          minLength: 3
          maxLength: 500
        sku:
          type: string
          minLength: 3
          maxLength: 60
        category:
          type: string
          minLength: 2
          maxLength: 80
        subcategory:
          type: string
          maxLength: 80
        brand:
          type: string
          minLength: 2
          maxLength: 80
        unitPrice:
          type: number
          format: double
          minimum: 0
        costPrice:
          type: number
          format: double
          minimum: 0
        stockQuantity:
          type: integer
          minimum: 0
        minimumStock:
          type: integer
          minimum: 0
    UpdatePartRequest:
      allOf:
        - $ref: '#/components/schemas/CreatePartRequest'
        - type: object
          required:
            - active
          properties:
            active:
              type: boolean
    UpdatePartStockRequest:
      type: object
      required:
        - stockQuantity
      properties:
        stockQuantity:
          type: integer
          minimum: 0
    StockMovementRequest:
      type: object
      required:
        - type
        - quantity
      properties:
        type:
          type: string
          enum:
            - ENTRY
            - EXIT
            - SALE
        quantity:
          type: integer
          minimum: 1
        unitCost:
          type: number
          format: double
          minimum: 0
        unitPrice:
          type: number
          format: double
          minimum: 0
        reason:
          type: string
          maxLength: 255
    ConfigurePartReservationRequest:
      type: object
      required:
        - reservationDays
      properties:
        reservationDays:
          type: integer
          minimum: 1
    PartQuantityRequest:
      type: object
      required:
        - quantity
      properties:
        quantity:
          type: integer
          minimum: 1
    CommitPartReservationRequest:
      allOf:
        - $ref: '#/components/schemas/PartQuantityRequest'
        - type: object
          properties:
            reason:
              type: string
              maxLength: 255
    PartResponse:
      type: object
      required:
        - id
        - name
        - description
        - sku
        - category
        - brand
        - unitPrice
        - costPrice
        - stockQuantity
        - reservedQuantity
        - minimumStock
        - stockStatus
        - active
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        sku:
          type: string
        category:
          type: string
        subcategory:
          type: string
        brand:
          type: string
        unitPrice:
          type: number
          format: double
        costPrice:
          type: number
          format: double
        stockQuantity:
          type: integer
        reservedQuantity:
          type: integer
        availableQuantity:
          type: integer
        minimumStock:
          type: integer
        reservationDays:
          type: integer
        reservationExpiresAt:
          type: string
          format: date-time
        stockStatus:
          type: string
        active:
          type: boolean
    PartListResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PartResponse'
...
    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          minimum: 100
          maximum: 599
        error:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
        path:
          type: string
          minLength: 1
        details:
          type: array
          items:
            type: string